rename main -> cli
