+++ tqdm/cli.py
